server/requirements.txt
# SAP Enterprise Decision Support AI — Python Dependencies

# Web Framework
fastapi==0.115.4
uvicorn[standard]==0.32.0

# AI / LLM
openai==1.54.0          # For OpenAI GPT-4o
anthropic==0.39.0       # For Claude (optional)

# Data / ML (for future real prediction models)
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2

# Config
python-dotenv==1.0.1
pydantic==2.9.2

# HTTP
httpx==0.27.2
requests==2.32.3

# Utils
python-multipart==0.0.17  # for file uploads
